changed my values to double instead of decimal

+++ Calculator/Mathematics/MathContext.cs
@@ -17,7 +17,7 @@ namespace Calculator.Mathematics
             _strategy = strategy;
         }
 
-        public decimal Calculate(decimal first, decimal second)
+        public double Calculate(double first, double second)
         {
             return _strategy.Execute(first, second);
         }
@@ -45,7 +45,6 @@ namespace Calculator.Mathematics
                     _strategy = new SquareRoot();
                     break;
             }
-
             return _strategy;
         }
     }
