+++ Calculator/Mathematics/Operations/Division.cs
@@ -10,7 +10,7 @@ namespace Calculator.Mathematics.Operations
     public class Division : IMathStrategy
     {
         public char Operator { get; set; } = '÷';
-        public decimal Execute(decimal first, decimal second)
+        public double Execute(double first, double second)
         {
             return Math.Round(first / second, 2);
         }
